query cancellation: terminate query if client disconnects
A close channel is attached to every task instance, so we can send a
message if the client hangs up. In the HTTP handler the context done
channel is used which which propagates that cancellation to the
worker. The worker uses db.QueryContext which aborts the query when the
context is canceled.

For the response channel and the close channel buffered channels (size
1) are used, so we don't accidentally lock up workers in case of
cancellation.

Resolves #6